Fix Painter's Palette dark slot and scrolling.

+++ src/client/java/dev/lambdaurora/aurorascanvas/client/mixin/MouseHandlerMixin.java
@@ -39,7 +39,7 @@ public class MouseHandlerMixin {
 	)
 	private void onScroll(
 			long window, double scrollDeltaX, double scrollDeltaY, CallbackInfo ci,
-			@Local(ordinal = 2) double scrollDelta
+			@Local(ordinal = 2) int scrollDelta
 	) {
 		var player = this.minecraft.player;
